Prep and pollish
Started preparing things for death and kill counters and actual death instead of being able to take damage for ever. Also polished weapon switching a bit. All that is left to do for that is to make it when you get sent the new loadout information it updates the gun model in your hand.