Created a buffer to manager matches
Created a separate C# command line application to handle the managing of matches and connecting clients too them. Also tweaked how the server and client work in order to use it. Took me about a week but Its done now. Next step is to finish the base game and add some sort of lobby system