my head is clearer than ever and its not fucked anymore :D
so for my future self: the movement is not smooth, it needs interpolation, no idea how that will work but it *might* not be that bad. guns and damage are yet to be implemented on the client and i havent tested this with multiple people. gravtiy is kinda off cause the server only sends a new state after a new input. also you should make it so that the forward of the player dosent initiate an input to the server it is just either bundled with that uses it or with regular updates to keep stuff clean. wont work for the future cause u kinda want to see people looking around but for now it dosent make sense that looking up and down can fix things. collision is bad cause the 'new' movement is just teliporting the player around. it will push you out because of the collider on smaller things but its not great. make the respawn thing stop you from moving at all, including gravity so it dosent keep that velocity when you spawn back in. if i find anything else ill add it to the list on ur whiteboard :thumbsup: